Professor refactor
